requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart>=0.0.9

# Gradio UI
gradio>=4.44.1,<5.0.0

# LangChain & LangGraph
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph==0.6.0
langgraph-checkpoint==2.1.1
langsmith==0.4.8

# Database
sqlalchemy==2.0.42
aiosqlite==0.19.0
supabase==2.10.0

# Environment & Configuration
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1

# Twilio (WhatsApp Integration)
twilio==8.13.0

# HTTP Requests
requests==2.32.4
httpx==0.28.1

# ChromaDB for RAG
chromadb==0.5.23

# Document Processing
pypdf==5.5.0
python-docx==1.1.0
# unstructured removed - not used in code and has Python version conflicts

# Scheduler
apscheduler==3.10.4

# OpenAI
openai==1.97.1

# Audio Processing
pydub==0.25.1

# Additional Dependencies
typing-extensions==4.12.2
packaging==24.1
